requirements.txt
pyyaml>=6.0
pyserial>=3.5